--- ADictionaryEcho/app/src/main/java/ayds/dictionary/echo/model/ServiceDefinitionInterface.java
@@ -1,5 +1,0 @@
-package ayds.dictionary.echo.model;
-
-public interface ServiceDefinitionInterface {
-    public String getTerm();
-}
