--- ADictionaryEcho/app/src/main/java/ayds/dictionary/echo/model/ServiceWikiAdapter.java
@@ -1,5 +1,0 @@
-package ayds.dictionary.echo.model;
-
-public class ServiceWikiAdapter implements ServiceDefinitionInterface {
-    public String getTerm(){return null;}
-}
