--- ADictionaryEcho/app/src/main/java/ayds/dictionary/echo/model/ServiceYandexAdapter.java
@@ -1,5 +1,0 @@
-package ayds.dictionary.echo.model;
-
-public class ServiceYandexAdapter implements ServiceDefinitionInterface {
-    public String getTerm(){return null;}
-}
