+++ ADictionaryEcho/app/src/main/java/ayds/dictionary/echo/model/business/services/ServiceBigHugLabsAdapter.java
@@ -1,0 +1,5 @@
+package ayds.dictionary.echo.model.business.services;
+
+public class ServiceBigHugLabsAdapter implements ServiceDefinitionInterface {
+    public String getTerm(){return null;}
+}
