+++ ADictionaryEcho/app/src/main/java/ayds/dictionary/echo/model/business/services/ServiceDefinitionInterface.java
@@ -1,0 +1,5 @@
+package ayds.dictionary.echo.model.business.services;
+
+public interface ServiceDefinitionInterface {
+    public String getTerm();
+}
