+++ ADictionaryEcho/app/src/main/java/ayds/dictionary/echo/model/business/services/ServiceYandexAdapter.java
@@ -1,0 +1,5 @@
+package ayds.dictionary.echo.model.business.services;
+
+public class ServiceYandexAdapter implements ServiceDefinitionInterface {
+    public String getTerm(){return null;}
+}
